feat: add Purchase, PurchaseDetail entity, add createPurchase method fix: repository annotation refer to issue #12

+++ src/main/java/com/example/sales_system/dto/response/ProductResponse.java
@@ -29,5 +29,5 @@ public class ProductResponse {
     String updateTime;
 
     @Builder.Default
-    Integer stock = 0;
+    Long stock = 0L;
 }
